refactor: move SpinalSimBmbLengthFixerTester
