+++ lib/src/test/scala/spinal/lib/bus/bmb/BmbSpecificBridges.scala
@@ -1,6 +1,7 @@
 package spinal.lib.bus.bmb
 
 import org.scalatest.funsuite.AnyFunSuite
+import spinal.tester.SpinalSimFunSuite
 
 import spinal.core._
 import spinal.core.sim.SimConfig
@@ -117,4 +118,91 @@ class SpinalSimBmbAlignerTester extends AnyFunSuite {
       }
     }
   }
+}
+
+class SpinalSimBmbLengthFixerTester extends SpinalSimFunSuite {
+  test("bypass") {
+    SimConfig.compile {
+      val c = BmbLengthFixer(
+        ip = BmbParameter(
+          addressWidth = 16,
+          dataWidth = 32,
+          lengthWidth = 6,
+          sourceWidth = 4,
+          contextWidth = 3,
+          alignmentMin = 2,
+          canRead = true,
+          canWrite = true,
+          alignment = BmbParameter.BurstAlignement.WORD
+        ),
+        fixedWidth = 2
+      )
+      c
+    }.doSimUntilVoid("test") { dut =>
+      new BmbBridgeTester(
+        master = dut.io.input,
+        masterCd = dut.clockDomain,
+        slave = dut.io.output,
+        slaveCd = dut.clockDomain,
+        alignmentMinWidth = dut.ip.access.alignmentMin
+      )
+    }
+  }
+
+  test("3") {
+    SimConfig.compile {
+      val c = BmbLengthFixer(
+        ip = BmbParameter(
+          addressWidth = 16,
+          dataWidth = 32,
+          lengthWidth = 6,
+          sourceWidth = 4,
+          contextWidth = 3,
+          alignmentMin = 3,
+          canRead = true,
+          canWrite = true,
+          alignment = BmbParameter.BurstAlignement.WORD
+        ),
+        fixedWidth = 3
+      )
+      c
+    }.doSimUntilVoid("test") { dut =>
+      new BmbBridgeTester(
+        master = dut.io.input,
+        masterCd = dut.clockDomain,
+        slave = dut.io.output,
+        slaveCd = dut.clockDomain,
+        alignmentMinWidth = dut.ip.access.alignmentMin
+      )
+    }
+  }
+
+  test("4") {
+    SimConfig.compile {
+      val c = BmbLengthFixer(
+        ip = BmbParameter(
+          addressWidth = 16,
+          dataWidth = 32,
+          lengthWidth = 6,
+          sourceWidth = 4,
+          contextWidth = 3,
+          alignmentMin = 4,
+          canRead = true,
+          canWrite = true,
+          alignment = BmbParameter.BurstAlignement.WORD
+        ),
+        fixedWidth = 4
+      )
+      c
+    }.doSimUntilVoid("test") { dut =>
+      new BmbBridgeTester(
+        master = dut.io.input,
+        masterCd = dut.clockDomain,
+        slave = dut.io.output,
+        slaveCd = dut.clockDomain,
+        alignmentMinWidth = dut.ip.access.alignmentMin
+      )
+    }
+  }
+
 }
